% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageSE.R
\name{averageSE}
\alias{averageSE}
\title{Average replicated samples in a SummarizedExperiment}
\usage{
averageSE(x, column = character(), mainAssay = character())
}
\arguments{
\item{x}{\code{SummarizedExperiment}.}

\item{column}{\code{character(1)} defining the column in \code{colData} that defines
replicates.}

\item{mainAssay}{optional \code{character(1)} defining an assay on which per-row
standard deviations across replicates should be calculated. The result
will be stored in an additional assay with the name of \code{mainAssay} and
suffix \code{"_sd"}.}
}
\value{
The averages \code{SummarizedExperiment}.
}
\description{
\code{averageSE} averages replicated samples (columns) in an
\code{SummarizedExperiment}.
}
\author{
Johannes Rainer
}
