% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatograms.R
\name{c,Chromatograms-method}
\alias{c,Chromatograms-method}
\alias{c,XChromatograms-method}
\title{Combine Chromatograms objects}
\usage{
\S4method{c}{Chromatograms}(x, ...)

\S4method{c}{XChromatograms}(x, ...)
}
\arguments{
\item{x}{\code{Chromatograms} object.}

\item{...}{\code{Chromatograms} object that should be appended to \code{x}.}
}
\value{
\code{Chromatograms} object.
}
\description{
\code{\link[=Chromatograms]{Chromatograms()}} objects can be concatenated (row-wise) with the \code{c}
function resulting in a \code{Chromatograms} object with the same number of
samples (columns).
}
\note{
The \code{c} function for \code{XChromatograms} objects (defined in the \code{xcms} package)
is not supported.
}
\author{
Johannes Rainer
}
