% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromPeakArea.R
\name{chromPeakArea}
\alias{chromPeakArea}
\title{Get MS peak area (m/z and rt range) for chromatographic peaks}
\usage{
chromPeakArea(
  x,
  mz = numeric(),
  rt = numeric(),
  diffRt = 40,
  ppm = 50,
  peakId = character()
)
}
\arguments{
\item{x}{\code{XCMSnExp} object with identified chromatographic peaks.}

\item{mz}{\code{numeric} or \code{matrix} with m/z values to identify chromatographic
peaks from which regions should be extracted. If a \code{matrix} is provided
it is expeted to have 2 columns with the lower and upper m/z boundary. If
a \code{numeric} is provided parameter \code{ppm} is also considered.}

\item{rt}{\code{numeric} or \code{matrix} with retention time values to identify
chromatographic peaks from which the peak areas are supposed to be
defined. If a \code{matrix} is provided it is expected to be a two column
matrix with the lower and upper rt boundary in which chromatographic
peaks should be found. If a \code{numeric} is provided, parameter \code{diffRt}
is considered to define the rt region to search for peaks.}

\item{diffRt}{\code{numeric(1)} to define the rt region in which peaks should be
identified if a \code{numeric} is provided with parameter \code{rt}. The rt region
is defined as \code{rt - diffRt} to \code{rt + diffRt}.}

\item{ppm}{\code{numeric(1)} to define the m/z region in which peaks should be
identified if \code{mz} is a \code{numeric}. The m/z region is defined as
\code{mz - ppm(mz, ppm)} to \code{mz + ppm(mz, ppm)}}
}
\value{
\code{matrix} with columnd \code{"mzmin"}, \code{"mzmax"}, \code{"rtmin"}, \code{"rtmax"}
with the m/z and retention time ranges calculated on all chromatographic
peaks in \code{x} matching \code{mz} and \code{rt}. Each row representing the result for
each value in \code{mz} and \code{rt}.
}
\description{
Extract the peak area (defined by the m/z and retention time range) of
chromatographic peaks. The chromatographic peaks can be defined by their ID
(i.e. rownames of the \code{chromPeaks} matrix) or by m/z and a retention times.

The function identifies first chromatographic peaks in \code{x} that match the
input parameters \code{mz} and \code{rt} or \code{peakId} and then defines the m/z and rt
range based on the \code{"mzmin"}, \code{"mzmax"}, \code{"rtmin"} and \code{"rtmax"} of these.
If m/z and rt ranges are provided, only chromatographic peaks are considered
that have their apex within the region.

As a result a \code{matrix} is returned with columns \code{"mzmin"}, \code{"mzmax"},
\code{"rtmin"} and \code{"rtmax"} defining the lower and upper boundaries of the
region of the peaks. Each row in this matrix represent the result for one
element in the input parameter specifying the peaks of interest (e.g. \code{mz},
\code{rt} or \code{peakId}).
}
\examples{

library(faahKO)
fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
         system.file('cdf/KO/ko16.CDF', package = "faahKO"),
         system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")
xod <- findChromPeaks(
   od, param = CentWaveParam(noise = 10000, snthresh = 40,
                             prefilter = c(3, 10000)))

chromPeakArea(xod, rt = c(2500, 2700), diffRt = 100,
    mz = cbind(c(400, 300), c(500, 400)))
}
\author{
Johannes Rainer
}
