% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropModels.R
\name{dropModels}
\alias{dropModels}
\title{Remove model fits based on specified criteria}
\usage{
dropModels(x, FLAG_FUN, ...)
}
\arguments{
\item{x}{\code{list} of linear models such as returned by \code{xcms:::rowFitModel} or
\code{\link[=withinBatchFit]{withinBatchFit()}}.}

\item{FLAG_FUN}{\code{function} to test whether the model fit is problematic or not,
such as \code{\link[=flag_model_residual]{flag_model_residual()}}. The funciton has to take a linear model
object (as generated by \code{\link[=lm]{lm()}}) as input and return \code{TRUE} if the model
should be removed or \code{FALSE} otherwise.}

\item{...}{additional parameters to be passed to \code{FUN}.}
}
\value{
The input \code{x} with poor quality models identified by \code{FUN} replaced
with \code{NA}.
}
\description{
The \code{dropModels} function takes model fits such as the ones returned by
\code{xcms:::rowFitModel} or \code{\link[=withinBatchFit]{withinBatchFit()}} and removes models idenfified by
function \code{FLAG_FUN}. This function can be one of the \verb{flag_*} functions (such
as \code{\link[=flag_model_residual]{flag_model_residual()}}) that take a linear model fit as input and return
a logical indicating whether the model fit is problematic or not.
}
\author{
Johannes Rainer
}
