% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_file_utils.R
\name{extract_time_stamp}
\alias{extract_time_stamp}
\title{Extract run start time stamp}
\usage{
extract_time_stamp(
  x = character(),
  format = "\%Y-\%m-\%dT\%H:\%M:\%S",
  BPPARAM = bpparam()
)
}
\arguments{
\item{x}{\code{character} with the file names (full path) or the mzML files.}

\item{format}{\code{character(1)} defining the date/time format of the time
stamp. If \code{NULL} the time stamp will be returned as a \code{character}.}

\item{BPPARAM}{Parallel processing setup. See \code{\link[=bpparam]{bpparam()}} for details.}
}
\value{
\code{character} (or date/time) with the start time stamps.
}
\description{
Extract the start time stamps from (mzML) files.
}
\author{
Johannes Rainer
}
