% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_utils.R
\name{extractSpectraData}
\alias{extractSpectraData}
\title{Extract data from MSnbase objects for use in Spectra}
\usage{
extractSpectraData(x)
}
\arguments{
\item{x}{a \code{list} of \link[MSnbase:Spectrum-class]{MSnbase::Spectrum} objects or an object extending
\link[MSnbase:MSnExp-class]{MSnbase::MSnExp} (such as also the \link[xcms:XCMSnExp-class]{xcms::XCMSnExp} object) or a
\link[MSnbase:MSpectra]{MSnbase::MSpectra} object.}
}
\value{
\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame} with the full spectrum data that can be
passed along to the \code{\link[Spectra:Spectra]{Spectra::Spectra()}} function to create such an
object.
}
\description{
\code{extractSpectraData} extracts the spectra data (m/z and intensity values
including metadata) from \link[MSnbase:Spectrum1-class]{MSnbase::Spectrum1} or \link[MSnbase:Spectrum2-class]{MSnbase::Spectrum2} objects
(or \code{list} of such objects) and returns these as a \code{DataFrame} that can
be used to create a \link[Spectra:Spectra]{Spectra::Spectra} object. This function enables thus
to convert data from the \emph{old} \code{MSnbase} package to the newer \code{Spectra}
package.
}
\examples{

## Read an mzML file with MSnbase
fl <- system.file("TripleTOF-SWATH", "PestMix1_SWATH.mzML",
    package = "msdata")
data <- MSnbase::filterRt(MSnbase::readMSData(fl, mode = "onDisk"),
    rt = c(1, 6))

## Extract the data as a DataFrame
res <- extractSpectraData(data)

## This can be used as an input for the Spectra constructor of the
## Spectra package.
sps <- Spectra::Spectra(res)
sps
}
\author{
Johannes Rainer
}
