% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_feature_methods.R
\name{feature-grouping}
\alias{feature-grouping}
\alias{featureGroups}
\title{Grouping of features}
\usage{
featureGroups(object)
}
\arguments{
\item{object}{an \code{\link[=XCMSnExp]{XCMSnExp()}} object.}
}
\description{
After correspondence analysis (\code{\link[xcms:groupChromPeaks]{xcms::groupChromPeaks()}}) the identified
features can be further grouped based on different criteria into
\emph{feature groups} which can ideally group features representing signals from
ions (adducts, isotopes) of the same compound/metabolite.

The available options for the \code{groupFeatures} method are:
\itemize{
\item Grouping by similar retention times: \code{\link[=SimilarRtimeParam]{SimilarRtimeParam()}}.
\item Grouping by similar feature values across samples:
\code{\link[=AbundanceCorrelationParam]{AbundanceCorrelationParam()}}.
\item Grouping by similar peak shape of extracted ion chromatograms:
\code{\link[=EicCorrelationParam]{EicCorrelationParam()}}.
}

An ideal workflow grouping features should sequentially perform the above
methods (in the listed order).
}
\author{
Johannes Rainer, Mar Garcia-Aloy, Vinicius Veri Hernandes
}
