% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_feature_methods.R
\name{groupFeatures-abundance-correlation}
\alias{groupFeatures-abundance-correlation}
\alias{AbundanceCorrelationParam}
\alias{groupFeatures,XCMSnExp,AbundanceCorrelationParam-method}
\title{Group features based on correlation of abundances across samples}
\usage{
AbundanceCorrelationParam(
  threshold = 0.9,
  value = "into",
  method = c("maxint", "medret", "sum"),
  intensity = "maxo",
  filled = TRUE,
  subset = integer(),
  greedy = FALSE,
  transform = log2
)

\S4method{groupFeatures}{XCMSnExp,AbundanceCorrelationParam}(object, param, msLevel = 1L)
}
\arguments{
\item{threshold}{\code{numeric(1)} with the minimal required correlation
coefficient to group featues.}

\item{value}{\code{character(1)} passed to the \code{featureValues} call. See
\code{\link[=featureValues]{featureValues()}} for details. Defaults to \code{value = "into"}.}

\item{method}{\code{character(1)} passed to the \code{featureValues} call. See
\code{\link[=featureValues]{featureValues()}} for details. Defaults to \code{method = "maxint"}.}

\item{intensity}{\code{character(1)} passed to the \code{featureValues} call. See
\code{\link[=featureValues]{featureValues()}} for details. Defaults to \code{intensity = "maxo"}.}

\item{filled}{\code{logical(1)} whether filled-in values should be included in
the correlation analysis. Defaults to \code{filled = TRUE}.}

\item{subset}{\code{integer} or \code{logical} defining a subset of samples (at least
2) on which the correlation should be performed.}

\item{greedy}{\code{logical(1)} which grouping algorithm should be used: one that
creates small groups of highly correlated features (\code{greedy = FALSE}, the
default) or whether features should be grouped that have at least one
correlation with any other member of the group in common
(\code{greedy = TRUE}). See \code{\link[=groupByCorrelation]{groupByCorrelation()}} for details.}

\item{transform}{\code{function} to be applied to the feature intensities prior
correlation. Defaults to \code{transform = log2}, i.e. log2 transforms the
data before correlation. To use the values \emph{as-is} use
\code{transform = identity}.}

\item{object}{\code{\link[=XCMSnExp]{XCMSnExp()}} object containing also correspondence results.}

\item{param}{\code{AbudanceCorrelationParam} object with the settings for the
method.}

\item{msLevel}{\code{integer(1)} defining the MS level on which the features
should be grouped.}
}
\value{
input \code{XCMSnExp} with feature groups added (i.e. in column
\code{"feature_group"} of its \code{featureDefinitions} data frame.
}
\description{
This methods groups features based on correlation of abundances (i.e.
\emph{feature values}) across samples. Parameter \code{subset} allows to define a sub
set of samples on which this correlation should be correlated. It might for
example be better to exclude QC samples from this correlation analysis
because values of all features are supposed to be constant in these samples
and including these could bias the correlation estimation. Also, it might
be better to perform the correlation including also \emph{gap filled} values which
is possible with parameter \code{filled = TRUE} on an \code{object} on which also
\code{\link[xcms:fillChromPeaks]{xcms::fillChromPeaks()}} was called.
}
\examples{

## Performing a quick preprocessing of a test data set.
library(faahKO)
fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
        system.file('cdf/KO/ko16.CDF', package = "faahKO"),
        system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")
xod <- findChromPeaks(
    od, param = CentWaveParam(noise = 10000, snthresh = 40,
                              prefilter = c(3, 10000)))
pdp <- PeakDensityParam(sampleGroups = c(1, 1, 2))
xodg <- groupChromPeaks(xod, param = pdp)

## Group features based on correlation of feature values (integrated
## peak area) across samples. Note that there are many missing values
## in the feature value which influence grouping of features in the present
## data set.
xodg_grp <- groupFeatures(xodg,
    param = AbundanceCorrelationParam(threshold = 0.8))
table(featureDefinitions(xodg_grp)$feature_group)

## Group based on the maximal peak intensity per feature
xodg_grp <- groupFeatures(xodg,
    param = AbundanceCorrelationParam(threshold = 0.8, value = "maxo"))
table(featureDefinitions(xodg_grp)$feature_group)
}
\seealso{
feature-grouping for a general overview.

Other feature grouping methods: 
\code{\link{groupFeatures-approximate-rtime}},
\code{\link{groupFeatures-eic-correlation}}
}
\author{
Johannes Rainer
}
\concept{feature grouping methods}
