% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_feature_methods.R
\name{groupFeatures-approximate-rtime}
\alias{groupFeatures-approximate-rtime}
\alias{SimilarRtimeParam}
\alias{groupFeatures,XCMSnExp,SimilarRtimeParam-method}
\title{Group features based on approximate retention times}
\usage{
SimilarRtimeParam(diffRt = 1, method = c("groupClosest", "greedy"))

\S4method{groupFeatures}{XCMSnExp,SimilarRtimeParam}(object, param, msLevel = 1L)
}
\arguments{
\item{diffRt}{\code{numeric(1)} defining the retention time window within which
features should be grouped. All features with a rtime difference
smaller or equal than \code{diffRt} are grouped.}

\item{method}{\code{character(1)} defining which grouping approach should be
taken. Allowed values are \code{method = "groupClosest"} (the default) and
\code{method = "greedy"}. See description for details.}

\item{object}{\code{\link[=XCMSnExp]{XCMSnExp()}} object containing also correspondence results.}

\item{param}{\code{SimilarRtimeParam} object with the settings for the method.}

\item{msLevel}{\code{integer(1)} defining the MS level on which the features
should be grouped.}
}
\value{
input \code{XCMSnExp} with feature groups added (i.e. in column
\code{"feature_group"} of its \code{featureDefinitions} data frame.
}
\description{
Group features based on similar retention time. This method is supposed to be
used as an initial \emph{crude} grouping of features based on the median retention
time of all their chromatographic peaks. All features with a difference in
their retention time which is \code{<=} parameter \code{diffRt} of the parameter object
are grouped together. If a column \code{"feature_group"} is found in
\code{\link[xcms:XCMSnExp-class]{xcms::featureDefinitions()}} this is further sub-grouped by this method.

Two different grouping methods are available:
\itemize{
\item \code{method = "greedy"}: this approach consecutively groups elements together
if their difference in retention time is smaller than \code{diffRt}. If two
features are grouped into one group, also all other features with a
retention time within the defined window to any of the two features are
also included into the feature group. This grouping is recursively
expanded which can lead, depending on \code{diffRt} to very large feature
groups spanning a large retention time window.
\item \code{method = "groupClosest"}: this approach uses the \code{\link[=groupClosest]{groupClosest()}} function
that groups values together if their difference is smaller than \code{diffRt}.
If the difference of a feature to more than one group is smaller \code{diffRt}
it is assigned to the group to which its retention time is closest (most
similar) to the mean retention time of that group. This leads to smaller
group sizes. See \code{\link[=groupClosest]{groupClosest()}} for details and examples.
}
}
\examples{

## Performing a quick preprocessing of a test data set.
library(faahKO)
fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
        system.file('cdf/KO/ko16.CDF', package = "faahKO"),
        system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")
xod <- findChromPeaks(
    od, param = CentWaveParam(noise = 10000, snthresh = 40,
                              prefilter = c(3, 10000)))
pdp <- PeakDensityParam(sampleGroups = c(1, 1, 2))
xodg <- groupChromPeaks(xod, param = pdp)

## Group features based on similar retention time (i.e. difference <= 2 seconds)
xodg_grp <- groupFeatures(xodg, param = SimilarRtimeParam(diffRt = 2))

## Feature grouping get added to the featureDefinitions in column "feature_group"
head(featureDefinitions(xodg_grp)$feature_group)

table(featureDefinitions(xodg_grp)$feature_group)
length(unique(featureDefinitions(xodg_grp)$feature_group))

## Using the "greedy" method to create larger groups
xodg_grp <- groupFeatures(xodg,
    param = SimilarRtimeParam(diffRt = 2, method = "greedy"))

length(unique(featureDefinitions(xodg_grp)$feature_group))
}
\seealso{
feature-grouping for a general overview.

Other feature grouping methods: 
\code{\link{groupFeatures-abundance-correlation}},
\code{\link{groupFeatures-eic-correlation}}
}
\author{
Johannes Rainer
}
\concept{feature grouping methods}
