% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_feature_methods.R
\name{groupFeatures-eic-correlation}
\alias{groupFeatures-eic-correlation}
\alias{EicCorrelationParam}
\alias{groupFeatures,XCMSnExp,EicCorrelationParam-method}
\title{Group features based on correlation of extracted ion chromatograms}
\usage{
EicCorrelationParam(
  threshold = 0.9,
  n = 1,
  clean = TRUE,
  value = c("maxo", "into"),
  greedy = FALSE
)

\S4method{groupFeatures}{XCMSnExp,EicCorrelationParam}(object, param, msLevel = 1L)
}
\arguments{
\item{threshold}{\code{numeric(1)} with the minimal required correlation
coefficient to group featues.}

\item{n}{\code{numeric(1)} defining the total number of samples per feature group
on which this correlation should be performed. This value is rounded up
to the next larger integer value.}

\item{clean}{\code{logical(1)} whether the correlation should be performed only
on the signals within the identified chromatographic peaks
(\code{clean = TRUE}, default) or all the signal from the extracted ion
chromatogram.}

\item{value}{\code{character(1)} defining whether samples should be grouped based
on the sum of the maximal peak intensity (\code{value = "maxo"}, the default)
or the integrated peak area (\code{value = "into"}) for a feature.}

\item{greedy}{\code{logical(1)} which grouping algorithm should be used: one that
creates small groups of highly correlated features (\code{greedy = FALSE}, the
default) or whether features should be grouped that have at least one
correlation with any other member of the group in common
(\code{greedy = TRUE}). See \code{\link[=groupByCorrelation]{groupByCorrelation()}} for more information.}

\item{object}{\code{\link[=XCMSnExp]{XCMSnExp()}} object containing also correspondence results.}

\item{param}{\code{EicCorrelationParam} object with the settings for the method.}

\item{msLevel}{\code{integer(1)} defining the MS level on which the features
should be grouped.}
}
\value{
input \code{XCMSnExp} with feature groups added (i.e. in column
\code{"feature_group"} of its \code{featureDefinitions} data frame.
}
\description{
Group features based on correlation of their extracted ion chromatograms.
This correlation is performed separately for each sample with the correlation
coefficients being aggregated across samples for the final comparison with
parameter \code{threshold} (the 75\% quantile of the per-sample correlation values
is used for the comparison with \code{threshold}).

This feature grouping should be called \strong{after} an initial feature
grouping by retention time (see \code{\link[=SimilarRtimeParam]{SimilarRtimeParam()}}). The feature groups
defined in columns \code{"feature_group"} of \code{featureDefinitions(object)} (for
features matching \code{msLevel}) will be used and refined by this method.
Features with a value of \code{NA} in \code{featureDefinitions(object)$feature_group}
will be skipped/not considered for feature grouping.

While being possible to be performed on the full data set without prior
feature grouping , this is not suggested for the following reasons: I) the
selection of the top \code{n} samples with the highest signal for the
\emph{feature group} will be biased by very abundant compounds as this is
performed on the full data set (i.e. the samples with the highest overall
intensities are used for correlation of all features) and II) it is
computationally much more expensive because a pairwise correlation between
all features has to be performed.

It is also suggested to perform the correlation on a subset of samples
per feature with the highest intensities of the peaks (for that feature)
although it would also be possible to run the correlation on all samples by
setting \code{n} equal to the total number of samples in the data set. EIC
correlation should however be performed ideally on samples in which the
original compound is highly abundant to avoid correlation of missing values
or noisy peak shapes as much as possible.

By default also the signal which is outside identified chromatographic peaks
is excluded from the correlation  (parameter \code{clean}).
}
\examples{

## Performing a quick preprocessing of a test data set.
library(faahKO)
fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
        system.file('cdf/KO/ko16.CDF', package = "faahKO"),
        system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")
xod <- findChromPeaks(
    od, param = CentWaveParam(noise = 10000, snthresh = 40,
                              prefilter = c(3, 10000)))
pdp <- PeakDensityParam(sampleGroups = c(1, 1, 2))
xodg <- groupChromPeaks(xod, param = pdp)

## Performing a feature grouping based on EIC correlation on a single
## sample
xodg_grp <- groupFeatures(xodg, param = EicCorrelationParam(n = 1))

table(featureDefinitions(xodg_grp)$feature_group)

## Usually it is better to perform this correlation on pre-grouped features
## e.g. based on approximate retention time.
xodg_grp <- groupFeatures(xodg, param = SimilarRtimeParam(diffRt = 4))
xodg_grp <- groupFeatures(xodg_grp, param = EicCorrelationParam(n = 1))

table(featureDefinitions(xodg_grp)$feature_group)
}
\seealso{
feature-grouping for a general overview.

Other feature grouping methods: 
\code{\link{groupFeatures-abundance-correlation}},
\code{\link{groupFeatures-approximate-rtime}}
}
\author{
Johannes Rainer
}
\concept{feature grouping methods}
