% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_feature_methods.R
\name{groupFeatures-similar-rtime}
\alias{groupFeatures-similar-rtime}
\alias{groupFeatures,XCMSnExp,SimilarRtimeParam-method}
\title{Group features based on similar retention times}
\usage{
\S4method{groupFeatures}{XCMSnExp,SimilarRtimeParam}(object, param, msLevel = 1L, ...)
}
\arguments{
\item{object}{\code{\link[=XCMSnExp]{XCMSnExp()}} object containing also correspondence results.}

\item{param}{\code{SimilarRtimeParam} object with the settings for the method. See
\code{\link[MsFeatures:groupFeatures-similar-rtime]{MsFeatures::SimilarRtimeParam()}} for details and options.}

\item{msLevel}{\code{integer(1)} defining the MS level on which the features
should be grouped.}

\item{...}{passed to the \code{groupFeatures} function on numeric values.}
}
\value{
input \code{XCMSnExp} with feature groups added (i.e. in column
\code{"feature_group"} of its \code{featureDefinitions} data frame.
}
\description{
Group features based on similar retention time. This method is supposed to be
used as an initial \emph{crude} grouping of features based on the median retention
time of all their chromatographic peaks. All features with a difference in
their retention time which is \code{<=} parameter \code{diffRt} of the parameter object
are grouped together. If a column \code{"feature_group"} is found in
\code{\link[xcms:XCMSnExp-class]{xcms::featureDefinitions()}} this is further sub-grouped by this method.

See \code{\link[MsFeatures:groupFeatures-similar-rtime]{MsFeatures::SimilarRtimeParam()}} in \code{MsFeatures} for more details.
}
\examples{

## Performing a quick preprocessing of a test data set.
library(faahKO)
fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
        system.file('cdf/KO/ko16.CDF', package = "faahKO"),
        system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")
xod <- findChromPeaks(
    od, param = CentWaveParam(noise = 10000, snthresh = 40,
                              prefilter = c(3, 10000)))
pdp <- PeakDensityParam(sampleGroups = c(1, 1, 2))
xodg <- groupChromPeaks(xod, param = pdp)

## Group features based on similar retention time (i.e. difference <= 2 seconds)
xodg_grp <- groupFeatures(xodg, param = SimilarRtimeParam(diffRt = 2))

## Feature grouping get added to the featureDefinitions in column "feature_group"
head(featureDefinitions(xodg_grp)$feature_group)

table(featureDefinitions(xodg_grp)$feature_group)
length(unique(featureDefinitions(xodg_grp)$feature_group))

## Using an alternative groupiing method that creates larger groups
xodg_grp <- groupFeatures(xodg,
    param = SimilarRtimeParam(diffRt = 2, groupFun = MsCoreUtils::group))

length(unique(featureDefinitions(xodg_grp)$feature_group))
}
\seealso{
Other feature grouping methods: 
\code{\link{groupFeatures-abundance-correlation}},
\code{\link{groupFeatures-eic-correlation}}
}
\author{
Johannes Rainer
}
\concept{feature grouping methods}
