% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{SimilarRtimeParam-class}
\alias{EicCorrelationParam-class}
\alias{AbundanceCorrelationParam-class}
\title{Internal page for hidden aliases}
\description{
Internal page for hidden aliases
}
\keyword{internal}
