% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joyPlot.R
\name{joyPlot}
\alias{joyPlot}
\alias{joyPlot,MChromatograms-method}
\alias{joyPlot,XCMSnExp-method}
\title{Create a stacked plot of multiple chromatograms}
\usage{
\S4method{joyPlot}{MChromatograms}(
  object,
  sample = 1,
  yoffset = 0.3,
  xlim = NULL,
  ylim = NULL,
  col = "#000000",
  peakBg = "#ffffffaa",
  type = "l",
  ylab = "",
  xlab = "retention time",
  spacing = c("relative", "equal"),
  legend = FALSE,
  ...
)

\S4method{joyPlot}{XCMSnExp}(
  object,
  sample = 1L,
  mz = c(-Inf, Inf),
  rt = c(-Inf, Inf),
  yoffset = 0.3,
  col = "#000000",
  peakBg = "#ffffffaa",
  type = "l",
  ylab = "",
  xlab = "retention time",
  spacing = c("relative", "equal"),
  legend = FALSE,
  ...
)
}
\arguments{
\item{object}{an \linkS4class{MChromatograms} or \linkS4class{XCMSnExp} object.}

\item{sample}{\code{integer(1)} defining the sample (column in \code{x}) from which the
plot should be created.}

\item{yoffset}{\code{numeric(1)} defining the proportion of the y-axis to be used
for the stacked chromatograms. The default value of \code{yoffset = 0.3} uses
30\% of the y axis to place the individual chromatograms (eiher equally
spaced or relative to their m/z). This value is relative to the largest
intensity value of all chromatograms.}

\item{xlim}{optional \code{numeric(2)} defining the x-axis limits.}

\item{ylim}{optional \code{numeric(2)} defining the y-axis limits.}

\item{col}{color of the lines used to draw the chromatograms. Can be of
length 1 or equal \code{nrow(x)}.}

\item{peakBg}{background color for the chromatograms. Can be of length 1
or equal \code{nrow(x)}.}

\item{type}{\code{character(1)} defining the plot type (see parameter \code{type} in
base \code{\link[=plot]{plot()}}.}

\item{ylab}{\code{character(1)} with the y-axis label.}

\item{xlab}{\code{character(1)} with the x-axis label.}

\item{spacing}{\code{character(1)} defining whether the chromatograms should be
placed on the y-axis relative to their m/z value (spacing = "relative")
or equally spaced (spacing = "equal").}

\item{legend}{\code{logical(1)} whether m/z values should be displayed next to
the chromatograms.}

\item{...}{additional arguments to be passed to the \code{plot} function.}

\item{mz}{for \code{object} being an \code{XCMSnExp}: \code{numeric(2)} defining the lower
and upper m/z of the m/z range.}

\item{rt}{for \code{object} being an \code{XCMSnExp}: \code{numeric(2)} defining the
retention time range.}
}
\value{
besides creating the plot the function returns (invisibly) the
y positions of the individual chromatograms.
}
\description{
The \code{joyPlot} function creates a stacked plot of chromatograms of a
single sample (parameter \code{sample}). The name \code{joyPlot} was chosen as the
created plot can, depending on the data, look similar to the cover of the
\href{https://en.wikipedia.org/wiki/Unknown_Pleasures}{Unknown Pleasures} album
of the band Joy Division.

The method is implemented for \linkS4class{XCMSnExp} and \linkS4class{MChromatograms}
objects:
\itemize{
\item For \code{object} being an \linkS4class{MChromatograms}: the function plots all
chromatograms stacked (overlappingly) on top of each others assuming that
the chromatograms are ordered by m/z (i.e. first row having the smallest,
last row the largest m/z).
\item For \code{object} being an \linkS4class{XCMSnExp}: the function will plot only
chromatograms for chromatographic peaks present in the specified m/z range
(parameter \code{mz}) and retention time range (parameter \code{rt}) in the sample
\code{sample}.
}
}
\examples{

library(xcms)

## Load a test data set with detected peaks
data(faahko_sub)
## Update the path to the files for the local system
dirname(faahko_sub) <- system.file("cdf/KO", package = "faahKO")

one <- filterFile(faahko_sub, 1L)

## Plot stacked chromatograms for m/z slices containing chromatographic peaks
joyPlot(one, rt = c(4000, 4200))

## Plot the chromatograms equally spaced and showing also their m/z
res <- joyPlot(one, rt = c(4000, 4200), spacing = "equal", legend = TRUE)

## joyPlot returns the y position of the slices/chromatograms
res

## Disable transparency of individual peaks
joyPlot(one, rt = c(4000, 4200), spacing = "equal", peakBg = "#ffffff")

## To create a joy plot for the full data, not just the slices containing
## chromatographic peaks we have to first extract the respective
## chromatograms from the data. Below we define thus the m/z slices - with
## a rather large m/z range for each (i.e. of 1) in the m/z range 400 - 500
mzs <- seq(400, 500, by = 1)
chr <- chromatogram(one, mz = cbind(mzs[-length(mzs)], mzs[-1]), include = "none", rt = c(4000, 4200))

joyPlot(chr)

## And specyfying a different color for each chromatogram
joyPlot(chr, col = heat.colors(nrow(chr)))

## Finally, creating a plot that looks more like the album cover
par(bg = "#000000")
joyPlot(chr, col = "#ffffff", peakBg = "#000000aa", yoffset = 0.7)

}
\author{
Johannes Rainer
}
