% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matchRtMz}
\alias{matchRtMz}
\title{Match features based on their retention time and m/z values}
\usage{
matchRtMz(
  x,
  table,
  nomatch = NA_integer_,
  rt_tolerance = 2,
  tolerance = 0,
  ppm = 20,
  duplicates = c("closest", "keep"),
  mzcol = "mz",
  rtcol = "rt"
)
}
\arguments{
\item{x}{\code{data.frame}, \code{matrix} or \code{DataFrame} with feature definitions (i.e.
m/z and retentin times) that should be matched against features in
\code{table}.}

\item{table}{\code{data.frame}, \code{matrix} or \code{DataFrame} with feature definitions
to match features in \code{x} against.}

\item{nomatch}{value that should be returned if no match for a feature is
found.}

\item{rt_tolerance}{\code{numeric(1)} with the largest acceptable difference in
retention time.}

\item{tolerance}{\code{numeric(1)} with a constant acceptable difference of m/z
values for features to be considered matching.}

\item{ppm}{\code{numeric(1)} with a m/z-dependent relative acceptable difference
(in parts per million) of m/z values.}

\item{duplicates}{\code{character(1)} whether the best (\verb{duplicates = "closest", default) or all matches (}duplicates = "keep"`) shpuld be returned.}

\item{mzcol}{\code{character(1)} with the name of the column containing the m/z
ratios.}

\item{rtcol}{\code{character(1)} with the name of the column containing the
retention times.}
}
\value{
for \code{duplicates = "closest"}: \code{integer} of length equal to \code{nrow(x)}
with the index of the row in \code{table} that matches each row in \code{x}
(e.g. \code{c(3, 4)} means the first feature in \code{x} matches with the 3rd
feature in \code{table}.
For \code{duplicates = "keep"}: \code{list} of length equal to \code{nrow(x)} with
indices of \strong{all} rows in \code{table} that match each row in \code{x}.
}
\description{
\code{matchRtMz} matches elements (features) in \code{x} with elements in \code{table}
based on similarity of their m/z \strong{and} retention time. With parameter
\code{duplicates = "closest"}, the function return the index of the best match,
considering the m/z difference of features that are within the acceptable
retention time difference (defined by \code{rt_tolerance}). With
\code{duplicates = "keep"} the index of all matching rows in \code{table} are returned.
}
\note{
The function first finds features in \code{table} with a difference of retention
time which is smaller than \code{rt_tolerance} and matches these using the
\code{\link[=closest]{closest()}} function.
}
\examples{

x <- data.frame(mz = c(23.4, 45.6, 56.9, 76.5, 76.5, 76.5, 80.1),
    rt = c(12, 34, 59, 34, 67, 65, 67))

set.seed(123)
y <- rbind(x, x)
y$mz <- y$mz + rnorm(nrow(y), sd = 0.0002)
y$rt[1:nrow(x)] <- x$rt + 2
y <- y[order(y$mz), ]

matchRtMz(x, y)

## Keeping all matches
matchRtMz(x, y, duplicates = "keep")

## Lower ppm
matchRtMz(x, y, duplicates = "keep", ppm = 5)

## even lower
matchRtMz(x, y, duplicates = "keep", ppm = 2)

matchRtMz(x, y, ppm = 0)
}
\author{
Johannes Rainer
}
