% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_flagging.R
\name{flag_model_residual}
\alias{flag_model_residual}
\alias{flag_model_mean_residual}
\alias{flag_model_inj_range}
\alias{flag_model_cat_count}
\title{Functions to flag/exclude models}
\usage{
flag_model_residual(x, diff_residual = 1)

flag_model_mean_residual(x, cut_off = 0.5)

flag_model_inj_range(x, min_range = 1, column = "inj_idx")

flag_model_cat_count(x, variable, min_count = 4)
}
\arguments{
\item{x}{a linear model object such as generated by \code{\link[=lm]{lm()}} or \code{\link[=lmrob]{lmrob()}}.}

\item{diff_residual}{for \code{flag_model_residual}: \code{numeric(1)} defining the
cut-off to flag models with large residuals. \code{TRUE} is reported for all
models with a difference between the 25 and 75 percent percentile of
residuals larger than this value.}

\item{cut_off}{for \code{flag_model_mean_residual}: \code{numeric(1)} defining the
cut-off to flag models with large average residuals.}

\item{min_range}{for \code{flag_model_inj_range}: \code{numeric(1)} defining the
minimum range. This is an absolute value, not a percentage. Means, the
function compares the \code{diff(range(x$x[, column]))} with \code{min_range} and
flags models that don't fit that criteria (i.e. have a smaller range).}

\item{variable}{for \code{flag_model_cat_count}: \code{character(1)} with the name of
the categorical value. Should be one of \code{colnames(x$model)}.}

\item{min_count}{for \code{flag_model_cat_count}: \code{integer(1)} with the minimum
required number of values within each category.}

\item{columns}{for \code{flag_model_inj_range}: \code{character(1)} specifying the
column containing the injection index.}
}
\value{
\code{logical(1)}: \code{TRUE} if model fit is potentially problematic and
\code{FALSE} otherwise or \code{NA} if no model provided.
}
\description{
The functions listed on this help page (\verb{flag_*}) allow to flag/identify
potentially problematic model fits. The function were designed to be used
with the models estimating an injection-order-dependent signal drift seen
in LC-MS based untargeted metabolomics data. Such models can be fit with the
\code{rowFitModel} function of the \code{xcms} package. Functions are expected to
return \code{TRUE} for potentially problematic model fits and \code{FALSE} otherwise.

The functions are:
\itemize{
\item \code{flag_model_residual}: test whether the difference between the difference
between the 25 and 75\% quantile of residuals is larger than the user
defined value \code{diff_residual}. This function identifies model fits with
on average large deviations of the individual data points from the fitted
line.
\item \code{flag_model_mean_residual}: tests if the \code{mean} of the absolute residuals
is larger than the provided value.
\item \code{flag_model_inj_range}: tests if values on which the model was fitted
spans a minimum required injection index range. This requires \code{x} being
a model of the type \code{y ~ inj_idx}.
\item \code{flag_model_cat_count}: tests if the number of replicated measurents
(categories) for a categorical variable (e.g. batch) are larger than
\code{min_count}.
}
}
\author{
Johannes Rainer
}
