% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatograms.R
\name{plotOverlay}
\alias{plotOverlay}
\title{Overlay plots from multiple EICs}
\usage{
plotOverlay(x, col = "#00000060", type = "l", main = NULL, ...)
}
\arguments{
\item{x}{\code{\link[=Chromatograms]{Chromatograms()}} object with the EICs to draw.}

\item{col}{color representation. Should be either of length 1 (to use the
same color for each EIC) or equal to \code{nrow(x)} to use a different color
for each EIC (row in \code{x}).}

\item{type}{\code{character(1)} specifying the plot type (see help from base
\code{plot} function). \code{type = "l"} (the default) draws lines, \code{type = "p"}
would draw each intensity as a point etc.}

\item{main}{optional \code{character} with the title for the plot of each sample.
Can be either of length 1 or equal \code{ncol(x)}.}

\item{...}{additional parameters to be passed down to the \code{points} function.}
}
\description{
\code{plotOverlay} draws chromatographic peak data from multiple (different)
extracted ion chromatograms (EICs) into the same plot. This allows to
directly compare the peak shape of these EICs in the same sample. In
contrast to the \code{plot} function for \code{\link[=Chromatograms]{Chromatograms()}} object, which draws
the data from the same EIC across multiple samples in the same plot, this
function draws the different EICs from the same sample into the same plot.
}
\author{
Johannes Rainer
}
