% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd.R
\name{rsd}
\alias{rsd}
\alias{rowRsd}
\title{Calculate relative standard deviations

\code{rsd} and \code{rowRsd} are convenience functions to calculate the relative
standard deviation (i.e. coefficient of variation) of a numerical vector
or for rows of a numerical matrix, respectively.}
\usage{
rsd(x, na.rm = TRUE, mad = FALSE)

rowRsd(x, na.rm = TRUE, mad = FALSE)
}
\arguments{
\item{x}{for \code{rsd} a \code{numeric}, for \code{rowRsd} a numerical \code{matrix}.}

\item{na.rm}{\code{logical(1)} whether missing values (\code{NA}) should be removed
prior to the calculations.}

\item{mad}{\code{logical(1)} whether the \emph{Median Absolute Deviation} (MAD) should
be used instead of the standard deviation. This is suggested for
non-gaussian distributed data.}
}
\description{
Calculate relative standard deviations

\code{rsd} and \code{rowRsd} are convenience functions to calculate the relative
standard deviation (i.e. coefficient of variation) of a numerical vector
or for rows of a numerical matrix, respectively.
}
\examples{

a <- c(4.3, 4.5, 3.6, 5.3)
rsd(a)

A <- rbind(a, a, a)
rowRsd(A)
}
\author{
Johannes Rainer
}
