% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_files.R
\name{sync_files_local}
\alias{sync_files_local}
\alias{remove_local_files}
\title{Create local copies of files}
\usage{
sync_files_local(files, remote_path, local_path)

remove_local_files(files, local_path)
}
\arguments{
\item{files:}{\code{character} with the file names (including relative paths).}

\item{remote_path:}{\code{character(1)} with the remote location of the files.}

\item{local_path:}{\code{character(1)} local path where to store the files.}
}
\description{
Copy \code{files} from \code{remote_path} to \code{local_path}. Parameter \code{files} could e.g.
be the files defined in a phenodata text file.

This functions, along with the related \code{remove_local_files}, help to
manage local copies of e.g. mzML files of an experiment and keep them in sync
with a central file storage. A typical use case for these functions is:
mzML files are kept in a central storage. To setup or run a specific analysis
some of these files (those belonging to the experiment) should be copied to
a local copy to enable also off-line analyses. Copying files manually can be
cumbersome and error prone, this function thus helps to identify and copy
the required files. After an analysis local files might again be removed
using the \code{remove_local_files} function.
}
\examples{

##fls <- c("2018/2018_02/20180203_20000630_NEG.mzML",
##    "2018/2018_02/20180203_20000703_NEG.mzML",
##    "2018/2018_02/20180203_20000735_NEG.mzML")

##sync_files_local(fls, remote_path = "/Users/jo/data",
##    local_path = "/Users/jo/tmp")
}
\author{
Johannes Rainer
}
