% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinBatchAdjust.R
\name{withinBatchAdjust}
\alias{withinBatchAdjust}
\title{Apply per-batch model adjustment of feature abundances

\code{withinBatchAdjust} adjusts feature abundances within each batch based
on the model provided with parameter \code{models} (which are fitted using the
\code{\link[=withinBatchFit]{withinBatchFit()}} function.}
\usage{
withinBatchAdjust(
  x,
  models,
  batch = x$batch,
  assay = "norm",
  log.transform = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{SummarizedExperiment} with the feature abundances that should be
adjusted.}

\item{models}{\code{list} of per-batch feature-wise model fits as returned by
\code{\link[=withinBatchFit]{withinBatchFit()}}.}

\item{batch}{\code{factor} defining the batch assignment of samples in \code{x}.}

\item{assay}{\code{character(1)} with the name of the assay matrix in \code{x} that
should be adjusted. Defaults to \code{"norm"}.}

\item{log.transform}{\code{logical(1)} defining whether the model fit has been
performed in log scale and the adjustment is also to be performed in
log scale. Note that even if \code{log.transform} is \code{TRUE} adjusted feature
abundances are returned in natural scale.}

\item{...}{additional arguments to be passed to
\code{xcms:::applyModelAdjustment}.}
}
\value{
\code{SummarizedExperiment} (input object \code{x}) with the feature abundances
of assay \code{assay} adjusted based on the provided models \code{models}.
}
\description{
Apply per-batch model adjustment of feature abundances

\code{withinBatchAdjust} adjusts feature abundances within each batch based
on the model provided with parameter \code{models} (which are fitted using the
\code{\link[=withinBatchFit]{withinBatchFit()}} function.
}
\author{
Johannes Rainer
}
