% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinBatchFit.R
\name{withinBatchFit}
\alias{withinBatchFit}
\title{Fit a model separately to each batch in an experiment}
\usage{
withinBatchFit(
  x,
  batch = x$batch,
  assay = "norm",
  model = y ~ injection_idx,
  method = "lmrob",
  minVals = 6,
  log.transform = TRUE
)
}
\arguments{
\item{x}{\code{SummarizedExperiment} with the data on which the model should be
fitted (feature-wise). To estimate the injection order dependent signal
drift on only QC samples, 'x' should contain only the data for the QC
samples of the experiment. It has to contain feature abundances and all
required sample annotations (such as a variable for \emph{batch} and one for
the injection index).}

\item{batch}{\code{factor} defining the batches.}

\item{assay}{\code{character(1)} specifying the}

\item{model}{model formula describing the model that should be fitted to
the data within each batch.}

\item{method}{\code{character(1)} specifying the method that should be used to
fit the model. Can be either \code{"lmrob"} for robust regression or \code{"lm"}
for least squares regression.}

\item{minVals}{\code{integer(1)} with the minimum required data values for the
model to be fitted.}

\item{log.transform}{\code{logical(1)} whether feature abundances should be
\code{log2} transformed prior fitting of the model (default is
\code{log.transform = TRUE}).}
}
\value{
\code{list} of length equal to the number of batches (levels of \code{batch})
each representing the model fits for one batch. \code{res[[1]]} is thus the
result for the first batch and is a \code{list} of fitted models, one for
each feature.
}
\description{
\code{withinBatchFit} fits the model \code{model} to each row of feature abundances
(in the assay matrix specified with \code{assay}) of \code{x}. This allows to perform
a per-batch-separate estimation of injection order dependent signal drifts.
}
\author{
Johannes Rainer
}
