% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_features.R
\name{groupByCorrelation}
\alias{groupByCorrelation}
\title{Group rows in a matrix based on their correlation}
\usage{
groupByCorrelation(
  x,
  method = "pearson",
  use = "pairwise.complete.obs",
  threshold = 0.9
)
}
\arguments{
\item{x}{\code{numeric} \code{matrix} where rows should be grouped based on
correlation of their values across columns being larger than \code{threshold}.}

\item{method}{\code{character(1)} with the method to be used for correlation. See
\code{\link[=corr]{corr()}} for options.}

\item{use}{\code{character(1)} defining which values should be used for the
correlation. See \code{\link[=corr]{corr()}} for details.}

\item{threshold}{\code{numeric(1)} defining the cut of value above which
rows are considered to be correlated and hence grouped.}
}
\value{
\code{list} of indices of rows which are grouped based on their
correlation.
}
\description{
The \code{groupByCorrelation} allows to group rows in a numeric matrix based on
their correlation with each other. Note that if for example row 1 and 3 have
a correlation above the threshold and rows 3 and 5 too (but correlation
between 1 and 5 is below the threshold) all 3 are grouped into the same
group (i.e. rows 1, 3 \strong{and} 5).
}
\examples{

x <- rbind(
    c(1, 3, 2, 5),
    c(2, 6, 4, 7),
    c(1, 1, 3, 1),
    c(1, 3, 3, 6),
    c(0, 4, 3, 1))

groupByCorrelation(x)
}
\seealso{
Other grouping operations: 
\code{\link{groupEicCorrelation}()}
}
\author{
Johannes Rainer
}
\concept{grouping operations}
