% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_feature_methods.R
\name{groupFeatures-abundance-correlation}
\alias{groupFeatures-abundance-correlation}
\alias{groupFeatures,XCMSnExp,AbundanceSimilarityParam-method}
\title{Group features based on similarity of abundances across samples}
\usage{
\S4method{groupFeatures}{XCMSnExp,AbundanceSimilarityParam}(
  object,
  param,
  msLevel = 1L,
  method = c("medret", "maxint", "sum"),
  value = "into",
  intensity = "into",
  filled = TRUE,
  ...
)
}
\arguments{
\item{object}{\code{\link[=XCMSnExp]{XCMSnExp()}} object containing also correspondence results.}

\item{param}{\code{AbudanceSimilarityParam} object with the settings for the
method. See \code{\link[=AbundanceSimilarityParam]{AbundanceSimilarityParam()}} for details on the grouping
method and its parameters.}

\item{msLevel}{\code{integer(1)} defining the MS level on which the features
should be grouped.}

\item{method}{\code{character(1)} passed to the \code{featureValues} call. See
\code{\link[=featureValues]{featureValues()}} for details. Defaults to \code{method = "medret"}.}

\item{value}{\code{character(1)} passed to the \code{featureValues} call. See
\code{\link[=featureValues]{featureValues()}} for details. Defaults to \code{value = "into"}.}

\item{intensity}{\code{character(1)} passed to the \code{featureValues} call. See
\code{\link[=featureValues]{featureValues()}} for details. Defaults to \code{intensity = "into"}.}

\item{filled}{\code{logical(1)} whether filled-in values should be included in
the correlation analysis. Defaults to \code{filled = TRUE}.}

\item{...}{additional parameters passed to the \code{groupFeatures} method for
\code{matrix}.}

\item{peakId}{\code{character} with the IDs of the chromatographic peaks (i.e.
rownames of \code{chromPeaks(object)}).}
}
\value{
input \code{XCMSnExp} with feature group definitions added to a column
\code{"feature_group"} in its \code{featureDefinitions} data frame.
}
\description{
This method groups features based on similarity of abundances (i.e.
\emph{feature values}) across samples. See also \code{\link[=AbundanceSimilarityParam]{AbundanceSimilarityParam()}} for
additional information and details.

This help page lists parameters specific for \code{xcms} result objects (i.e. the
\code{\link[=XCMSnExp]{XCMSnExp()}} object). Documentation of the parameters for the similarity
calculation is available in the \code{\link[=AbundanceSimilarityParam]{AbundanceSimilarityParam()}} help page in
the \code{MsFeatures} package.
}
\examples{

## Performing a quick preprocessing of a test data set.
library(faahKO)
fls <- c(system.file('cdf/KO/ko15.CDF', package = "faahKO"),
        system.file('cdf/KO/ko16.CDF', package = "faahKO"),
        system.file('cdf/WT/wt19.CDF', package = "faahKO"))

od <- readMSData(fls, mode = "onDisk")
xod <- findChromPeaks(
    od, param = CentWaveParam(noise = 10000, snthresh = 40,
                              prefilter = c(3, 10000)))
pdp <- PeakDensityParam(sampleGroups = c(1, 1, 2))
xodg <- groupChromPeaks(xod, param = pdp)

## Group features based on correlation of feature values (integrated
## peak area) across samples. Note that there are many missing values
## in the feature value which influence grouping of features in the present
## data set.
xodg_grp <- groupFeatures(xodg,
    param = AbundanceSimilarityParam(threshold = 0.8))
table(featureDefinitions(xodg_grp)$feature_group)

## Group based on the maximal peak intensity per feature
xodg_grp <- groupFeatures(xodg,
    param = AbundanceSimilarityParam(threshold = 0.8, value = "maxo"))
table(featureDefinitions(xodg_grp)$feature_group)
}
\seealso{
feature-grouping for a general overview.

Other feature grouping methods: 
\code{\link{groupFeatures-eic-correlation}},
\code{\link{groupFeatures-similar-rtime}}
}
\author{
Johannes Rainer
}
\concept{feature grouping methods}
